Move ghi from cmd/ and make github-cli root command
